fix: share file database url parsing
